recipes-bsp: u-boot-phytec-imx: Fix for non OF_UPSTREAM boards
Not all PHYTEC i.MX8M boards have migrated to OF_UPSTREAM and hence
do_deploy task for imx-boot-tools dtb install fails. Implement a fix
that handles both situations (OF_UPSTREAM or in-tree dts).

+++ recipes-bsp/u-boot/u-boot-phytec-imx_2025.04-2.2.0-phy10.bb
@@ -33,7 +33,11 @@ do_deploy:append:mx8m-generic-bsp() {
                 if [ $j -eq $i ]
                 then
                     install -d ${DEPLOYDIR}/imx-boot-tools
-                    install -m 0777 ${B}/${config}/dts/upstream/src/arm64/freescale/${UBOOT_DTB_NAME} ${DEPLOYDIR}/imx-boot-tools
+                    if [ -f "${B}/${config}/dts/upstream/src/arm64/freescale/${UBOOT_DTB_NAME}" ]; then
+                        install -m 0777 ${B}/${config}/dts/upstream/src/arm64/freescale/${UBOOT_DTB_NAME} ${DEPLOYDIR}/imx-boot-tools
+                    else
+                        install -m 0777 ${B}/${config}/arch/arm/dts/${UBOOT_DTB_NAME} ${DEPLOYDIR}/imx-boot-tools
+                    fi
                     install -m 0777 ${B}/${config}/u-boot-nodtb.bin  ${DEPLOYDIR}/${BOOT_TOOLS}/u-boot-nodtb.bin-${MACHINE}-${type}
                 fi
             done
